pyproject.toml
[project]
name = "rss-digest"
version = "0.3.0"
description = "Content Curation Engine - fetch, process, and publish RSS digests with a modular two-phase architecture"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.11",
    "python-dateutil>=2.8.2",
    "openai>=1.12.0",
    "python-dotenv>=1.0.1",
    "toml>=0.10.2",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "markdown>=3.5.0",
]

[project.scripts]
rss-digest = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
